gcc compiler warnings and more 64bit progress

+++ trunk/Omnibot/Common/Waypoint.h
@@ -10,6 +10,7 @@
 #define __WAYPOINT_H__
 
 #include "NavigationFlags.h"
+#include <stdint.h>
 
 // class: Waypoint
 //		3d waypoint within the game for the waypoint path planner.
@@ -63,8 +64,8 @@ public:
 
 	inline void SetName( const String & s ) { m_WaypointName = s; m_NeedsSynced = true; }
 	inline const String &GetName() const { return m_WaypointName; }
-	inline int GetUID() const { return m_UID; }
-	inline int GetHash() const { return (unsigned int)this; }
+	inline int GetUID() const { return (int)m_UID; }
+	inline int GetHash() const { return (intptr_t)this; }
 	inline void SetEntity(GameEntity _ent) { m_Entity = _ent; }
 
 	bool IsConnectedTo(const Waypoint *_wp) const;
